version .5
code is for demonstration only. Not meant for public consumption. In
the near future I will be cleaning this up a lot and adding many
features to make it useable by a wide audience. This was built for a
production site and as an experiment to integrate jQuery
navigation/animation with knockout.js data binding.